requirements.txt
selenium
tqdm
mysql-connector-python